Added Wikipedia Traversal Bot; Started Optimization
Added a bot to the force renderer. Will traverse wikipedia in real time
from a starting point and add nodes and edges to the renderer as it
goes.

Has variables for limiting the number of pages and number of links
gathered per page, as well as filters to ensure only links to other
wikipedia articles are clicked

Began optimizing force directed renderer with a grid based system for
rapidly gathering node neighbors